Rename `src/poll.check.ts`
